Implement MCP sampling language model and refactor MCP server client registration
