+++ packages/cli/src/mcpserver.ts
@@ -4,9 +4,7 @@
 import {
   CHANGE,
   CORE_VERSION,
-  MODEL_PROVIDER_MCP,
   RESOURCE_CHANGE,
-  SYSTEM_FENCE,
   TOOL_ID,
   deleteUndefinedValues,
   ensureDotGenaiscriptPath,
@@ -14,21 +12,15 @@ import {
   genaiscriptDebug,
   logVerbose,
   logWarn,
-  parseModelIdentifier,
   runtimeHost,
   setConsoleColors,
   splitMarkdownTextImageParts,
   toStrictJSONSchema,
+  mcpCreateLanguageModel,
 } from "@genaiscript/core";
 import type {
-  ChatCompletionResponse,
-  ChatCompletionsOptions,
-  CreateChatCompletionRequest,
   GenerationResult,
   JSONSchemaObject,
-  LanguageModel,
-  LanguageModelConfiguration,
-  MarkdownTrace,
   Resource,
   ResourceContents,
   ScriptFilterOptions,
@@ -38,8 +30,6 @@ import {
   ListResourcesRequestSchema,
   ListResourceTemplatesRequestSchema,
   ReadResourceRequestSchema,
-  CreateMessageRequestSchema,
-  CreateMessageResultSchema,
 } from "@modelcontextprotocol/sdk/types.js";
 import type {
   CallToolResult,
@@ -51,9 +41,7 @@ import type {
 import { applyRemoteOptions } from "./remote.js";
 import type { RemoteOptions } from "./remote.js";
 import { startProjectWatcher } from "./watch.js";
-import { stderr } from "@genaiscript/core";
 const dbg = genaiscriptDebug("mcp:server");
-const dbgs = genaiscriptDebug("mcp:server:sampling");
 
 /**
  * Starts the MCP server.
@@ -167,7 +155,7 @@ export async function startMcpServer(
           } else {
             dbg(`unknown message type: %s`, data.type);
           }
-        }
+        },
       })) || { status: "error", error: { message: "run failed" } };
       dbg(`res: %s`, res.status);
       if (res.error) dbg(`error: %O`, res.error);
@@ -232,75 +220,7 @@ export async function startMcpServer(
     dbg(`client capabilities: %O`, clientCapabilities);
     if (clientCapabilities?.sampling) {
       dbg(`registering client sampling`);
-      runtimeHost.clientLanguageModel = Object.freeze<LanguageModel>({
-        id: MODEL_PROVIDER_MCP,
-        completer: async (
-          req: CreateChatCompletionRequest,
-          connection: LanguageModelConfiguration,
-          completerOptions: ChatCompletionsOptions,
-          trace: MarkdownTrace,
-        ): Promise<ChatCompletionResponse> => {
-          // Implement the completer logic here
-          dbgs(`sampling ${req.model}`);
-          const { model } = parseModelIdentifier(req.model);
-          const { partialCb, inner } = completerOptions || {};
-
-          const maxTokens = req.max_completion_tokens;
-          const systemMessages = req.messages.filter(({ role }) => role === "system");
-          const systemPrompt = systemMessages.map(({ content }) => content).join(SYSTEM_FENCE);
-          const otherMessages = req.messages.filter(({ role }) => role !== "system");
-
-          const body = deleteUndefinedValues({
-            method: "sampling/createMessage",
-            params: deleteUndefinedValues({
-              messages: otherMessages,
-              temperature: req.temperature,
-              metadata: req.metadata,
-              modelPreferences: {
-                hints: [
-                  {
-                    name: model,
-                  },
-                ].filter(({ name }) => !!name),
-                intelligencePriority: 0.8,
-                speedPriority: 0.5,
-              },
-              systemPrompt,
-              maxTokens,
-            }),
-          });
-
-          trace.detailsFenced(`🧪 mcp sampling`, body, "json");
-
-          let responseSoFar = "";
-          const res = await server.request(body, CreateMessageResultSchema, {
-            onprogress: (data) => {
-              dbgs(`%d/%d %s`, data.progress, data.total, data.message);
-              responseSoFar += data.message;
-              partialCb?.({
-                responseSoFar,
-                responseChunk: data.message,
-                tokensSoFar: data.progress,
-                inner,
-              });
-            },
-          });
-
-          trace.detailsFenced(`🧪 sampling result`, res, "json");
-          // "endTurn", "stopSequence", "maxTokens"
-          const finishReason: "stop" | "length" | "fail" =
-            {
-              ["endTurn"]: "stop",
-              ["stopSequence"]: "stop",
-              ["maxTokens"]: "length",
-            }[res.stopReason] ?? ("fail" as any);
-          return {
-            model: res.model,
-            text: res.content?.type === "text" ? res.content.text : "",
-            finishReason,
-          } satisfies ChatCompletionResponse;
-        },
-      } satisfies LanguageModel);
+      runtimeHost.clientLanguageModel = mcpCreateLanguageModel(server);
     }
   };
